Changing episode links to use an ObjectId instead of a base64'd string.
Episode links went off of the episode's _id, which was either a guid or the episode's link. There is the chance that this could not be unique (two links are set to be the same, or a magical guid collision). Since it wasn't necessarily unique and wasn't particularly nice or easy to work with, an ObjectId is a good alternative.