Changing error handling during scraping to no longer halt the entire scraping process when one source has an error.
If a source has an error (e.g. the RSS site times out) then the scraping process stopped completely, this has been changed in this commit to log the error and move on with the scraping as it does not necessarily make sense to stop all activities just because one didn't work. This could have been a bigger problem when the number of scraped sources increases to say the hundreds, could have run in to a situation where nothing gets scraped because of one site each time.